move helpers into `mosaicmllogger_utils.py`
