+++ scripts/train/train.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 import copy
 import gc
-import json
 import logging
 import os
 import sys
@@ -24,7 +23,8 @@ from omegaconf import DictConfig, ListConfig
 from omegaconf import OmegaConf as om
 from rich.traceback import install
 
-from llmfoundry.utils.logging_utils import get_cloud_provider_from_path
+from llmfoundry.utils.mosaicmllogger_utils import (find_mosaicml_logger,
+                                                   log_train_analytics)
 
 install()
 
@@ -151,107 +151,6 @@ def build_composer_model(model_cfg: DictConfig,
         raise ValueError(
             f'Not sure how to build model with name={model_cfg.name}')
     return COMPOSER_MODEL_REGISTRY[model_cfg.name](model_cfg, tokenizer)
-
-
-# TODO cleanup / decompose logic
-def log_analytics_details(mosaicml_logger: MosaicMLLogger,
-                          model_config: DictConfig,
-                          train_loader_config: DictConfig,
-                          eval_loader_config: Union[DictConfig, ListConfig,
-                                                    None],
-                          callback_configs: Union[DictConfig,
-                                                  None], tokenizer_name: str,
-                          load_path: Union[str, None], save_folder: Union[str,
-                                                                          None],
-                          icl_tasks_config: Optional[Union[ListConfig, str]],
-                          eval_gauntlet: Optional[Union[DictConfig, str]]):
-    train_loader_dataset = train_loader_config.get('dataset', {})
-    metrics: Dict[str, Any] = {
-        'llmfoundry/tokenizer_name':
-            tokenizer_name,
-        'llmfoundry/script':
-            'train',
-        'llmfoundry/train_loader_name':
-            train_loader_config.get('name'),
-        'llmfoundry/train_loader_workers':
-            train_loader_dataset.get('num_workers'),
-    }
-
-    if callback_configs is not None:
-        metrics['llmfoundry/callbacks'] = [
-            name for name, _ in callback_configs.items()
-        ]
-
-    if eval_gauntlet is not None:
-        metrics['llmfoundry/gauntlet_configured'] = True
-    else:
-        metrics['llmfoundry/gauntlet_configured'] = False
-
-    if icl_tasks_config is not None:
-        if isinstance(icl_tasks_config, str):
-            metrics['llmfoundry/icl_configured'] = True
-        elif len(icl_tasks_config) > 0:
-            metrics['llmfoundry/icl_configured'] = True
-        else:
-            metrics['llmfoundry/icl_configured'] = False
-    else:
-        metrics['llmfoundry/icl_configured'] = False
-
-    if load_path is not None:
-        metrics[
-            'llmfoundry/cloud_provider_data'] = get_cloud_provider_from_path(
-                load_path)
-    if save_folder is not None:
-        metrics[
-            'llmfoundry/cloud_provider_checkpoints'] = get_cloud_provider_from_path(
-                save_folder)
-
-    if train_loader_dataset.get('hf_name', None) is not None:
-        metrics['llmfoundry/train_dataset_hf_name'] = train_loader_dataset.get(
-            'hf_name', None)
-    if train_loader_config.get('name') == 'finetuning':
-        metrics['llmfoundry/train_task_type'] = 'INSTRUCTION_FINETUNE'
-    elif train_loader_config.get('name') == 'text':
-        if load_path is not None or model_config.get('pretrained') == True:
-            metrics['llmfoundry/train_task_type'] = 'CONTINUED_PRETRAIN'
-        else:
-            metrics['llmfoundry/train_task_type'] = 'PRETRAIN'
-
-    if eval_loader_config is not None:
-        metrics['llmfoundry/eval_loaders'] = []
-
-        if isinstance(eval_loader_config, ListConfig):
-            eval_loader_configs: ListConfig = eval_loader_config
-        else:
-            eval_loader_configs = ListConfig([eval_loader_config])
-
-        for loader_config in eval_loader_configs:
-            eval_loader_info = {}
-            eval_loader_dataset = loader_config.get('dataset', {})
-            eval_loader_info['name'] = loader_config.get('name')
-            eval_loader_info['num_workers'] = eval_loader_dataset.get(
-                'num_workers', None)
-            if eval_loader_dataset.get('hf_name', None) is not None:
-                eval_loader_info['dataset_hf_name'] = eval_loader_dataset.get(
-                    'hf_name')
-
-            # Log as a key-sorted JSON string, so that we can easily parse it in Spark / SQL
-            metrics['llmfoundry/eval_loaders'].append(
-                json.dumps(eval_loader_info, sort_keys=True))
-
-    # TODO: do we need error checking here?
-    if model_config['name'] == 'hf_casual_lm':
-        metrics['llmfoundry/model_name'] = model_config.get(
-            'pretrained_model_name_or_path')
-    if model_config.get('vocab_size', None) is not None:
-        metrics['llmfoundry/vocab_size'] = model_config.get('vocab_size'),
-    if model_config.get('d_model', None) is not None:
-        metrics['llmfoundry/d_model'] = model_config.get('d_model')
-    if model_config.get('n_heads', None) is not None:
-        metrics['llmfoundry/n_heads'] = model_config.get('n_heads')
-
-    mosaicml_logger.log_metrics(metrics)
-    mosaicml_logger._flush_metadata(force_flush=True)
 
 
 def main(cfg: DictConfig) -> Trainer:
@@ -553,9 +452,7 @@ def main(cfg: DictConfig) -> Trainer:
         for name, logger_cfg in logger_configs.items()
     ] if logger_configs else []
 
-    mosaicml_logger = next(
-        (logger for logger in loggers if isinstance(logger, MosaicMLLogger)),
-        None)
+    mosaicml_logger = find_mosaicml_logger(loggers)
     if mosaicml_logger is None:
         if os.environ.get(MOSAICML_PLATFORM_ENV_VAR, 'false').lower(
         ) == 'true' and os.environ.get(MOSAICML_ACCESS_TOKEN_ENV_VAR):
@@ -648,11 +545,10 @@ def main(cfg: DictConfig) -> Trainer:
             callbacks.append(eval_gauntlet_callback)
 
     if mosaicml_logger is not None:
-        log_analytics_details(mosaicml_logger, model_config,
-                              train_loader_config, eval_loader_config,
-                              callback_configs, tokenizer_name, load_path,
-                              save_folder, icl_tasks_config,
-                              eval_gauntlet_config)
+        log_train_analytics(mosaicml_logger, model_config, train_loader_config,
+                            eval_loader_config, callback_configs,
+                            tokenizer_name, load_path, save_folder,
+                            icl_tasks_config, eval_gauntlet_config)
 
     # Build Model
     log.info('Initializing model...')
